update the output of the code.
- For the first `console.log` and `process.nextTick()`, their output is okay. 

- But the output for setTimeout and setImmediate its not okay.

Here is the reason why

- Callbacks in the timer's queue ( callbacks set by setTimeout/setInterval ) are executed before callbacks in the check queue (those set by setImmediate). When setTimeout delay is set to 0ms, Nodejs interprets it as 1ms forexample

In case 1, if the event loop enters the timer's queue in 0.5ms that means that the callback has not been enqueued yet so the timer's queue will be empty therefore the callback will be executed on the second run. And we assume that their is not further user sync code to execute. therefore the output will be as follows:-
 
  1. Running last => number 4
  2. The timeout running before setImmediate => number 3

In case 2, if the CPU is busy and the event loop enter the timer's queue in 1.01ms, the 1ms will have elapsed implying that the callback will have been enqueued therefore the output will be as:- 

1. The timeout running before setImmediate => number 3
2. Running last => number 4

The above results in case 2 is what i got when i run the code on node v16.8.0


Signed-off-by: isaac <87535414+katongole-isaac@users.noreply.github.com>